showroom/usecase-002/requirements.txt
PyJWT>=2.8.0
cryptography>=44.0.0
requests>=2.32.0
python-dotenv>=1.0.0
flask>=2.3.0
pyngrok>=7.0.0